Data migration Validaiton Scripts commited

+++ DBQA  Validation  Script/reHabit@/Data Migration Bulk and Delta Validations/DELTA/Orders/uspOrderConsignmentNEntry_delta.sql
@@ -41,11 +41,12 @@ AS
 
 
         SET NOCOUNT ON 
-        SET @message = ' STEP: 1.RFOSource Table Started to Load.'
+        SET @message = CONCAT(' STEP: 1.RFOSource Table Started to Load.',
+                              CHAR(10),
+                              '-----------------------------------------------')
         EXECUTE dbqa.uspPrintMessage @message
 
-        SELECT  Quantity ,--	p_quantity
-                Quantity ,--	p_shippedquantity
+        SELECT  Quantity ,--	p_shippedquantity
                 TrackingNumber ,--	p_trackingid
                 ShipDate ,--	p_shippingdate
                 ro.OrderNumber AS [RFOKey] ,--	P_code
@@ -61,14 +62,15 @@ AS
         WHERE   EXISTS ( SELECT 1
                          FROM   Hybris.dbo.orders ho
                                 JOIN Hybris.dbo.Composedtypes ct ON ct.pk = ho.typepkstring
-                                                              AND ct.internalcode <> 'returnOrder' )
+                                                              AND ct.internalcode = 'Order' )
                 AND osp.ServerModifiedDate BETWEEN @StartDate
                                            AND     @EndDate
 
 
 
-
-        SET @message = 'STEP: 2.RFOSource Table Loaded and Target Table Started to Load'
+        SET @message = CONCAT('STEP: 2. Target Table Started to Load',
+                              CHAR(10),
+                              '-----------------------------------------------')
         EXECUTE dbqa.uspPrintMessage @message
 
 
@@ -85,14 +87,15 @@ AS
                 JOIN Hybris.dbo.consignmententries ce ON ce.p_consignment = c.PK
                 JOIN Hybris.dbo.orders ho ON ho.pk = c.p_order
                 JOIN Hybris.dbo.composedtypes ct ON ct.pk = ho.TypePkString
-                                                    AND ct.InternalCode <> 'returnOrder'
                 JOIN Hybris.dbo.enumerationvalues v ON v.pk = c.p_status
                 JOIN Hybris.dbo.deliverymodes m ON m.pk = c.p_deliverymode
-        WHERE   ( ( c.ModifiedTS BETWEEN @StartDate AND @EndDate )
-                  OR ( ce.ModifiedTS BETWEEN @StartDate AND @EndDate )
-                )
+        WHERE   ct.InternalCode = 'Order'
+                AND ( ( c.ModifiedTS BETWEEN @StartDate AND @EndDate )
+                      OR ( ce.ModifiedTS BETWEEN @StartDate AND @EndDate )
+                    )
 
-        SET @message = ' STEP:3.Hybris Table Loaded and Starting to Validate'
+        SET @message = CONCAT(' STEP:3.Creating Indexes in Temps', CHAR(10),
+                              '-----------------------------------------------')
         EXECUTE dbqa.uspPrintMessage @message     
 	
         CREATE NONCLUSTERED INDEX cls_RFO ON #RFO(RFOKey)  
@@ -104,7 +107,9 @@ AS
 -- TOTAL COUNT VALIDATION
 --++++++++++++++++++++++++++++++++++++
 
-        SET @message = ' STEP: 4.initiating COUNT Validation  '
+        SET @message = CONCAT(' STEP: 4.initiating COUNT Validation  ',
+                              CHAR(10),
+                              '-----------------------------------------------')
         EXECUTE dbqa.uspPrintMessage @message
 
         SELECT  @SourceCount = COUNT([RFOKey])
@@ -146,7 +151,9 @@ AS
 --++++++++++++++++++--++++++++++++++++++
 -- RFO  Duplicate Vaidation.
 --++++++++++++++++++--++++++++++++++++++
-        SET @message = ' STEP: 5.initiating DUPLICATE Need to updaete Scripts   '
+        SET @message = CONCAT(' STEP: 5.initiating DUPLICATE Need to updaete Scripts and Delete from EndToEnd  ',
+                              CHAR(10),
+                              '-----------------------------------------------')
         EXECUTE dbqa.uspPrintMessage @message
 
 --SET @SourceCount = 0
@@ -259,7 +266,9 @@ AS
 --++++++++++++++++++++++++++++++++++++++++++
 -- MISSING IN SOURCE AND MISSING IN TARGET 
 --++++++++++++++++++++++++++++++++++++++++++
-        SET @message = ' STEP: 6.initiating MISSING Validation now '
+        SET @message = CONCAT(' STEP: 6.initiating MISSING Validation ',
+                              CHAR(10),
+                              '-----------------------------------------------')
         EXECUTE dbqa.uspPrintMessage @message
  
         SELECT  a.[RFOKey] ,
@@ -336,7 +345,7 @@ AS
                   Owner ,
                   Flag ,
                   SourceColumn ,
-                  TargetCoulumn ,
+                  TargetColumn ,
                   [Key] ,
                   SourceValue ,
                   TargetValue
@@ -346,7 +355,7 @@ AS
                         @Owner , -- Owner - nvarchar(50)
                         [Missing From ] , -- Flag - nvarchar(10)
                         N'' , -- SourceColumn - nvarchar(50)
-                        N'' , -- TargetCoulumn - nvarchar(50)
+                        N'' , -- TargetColumn - nvarchar(50)
                         @Key , -- Key - nvarchar(50)
                         [RFOKey] , -- SourceValue - nvarchar(50)
                         [HybrisKey]
@@ -358,7 +367,7 @@ AS
                         @Owner , -- Owner - nvarchar(50)
                         [Missing From ] , -- Flag - nvarchar(10)
                         N'' , -- SourceColumn - nvarchar(50)
-                        N'' , -- TargetCoulumn - nvarchar(50)
+                        N'' , -- TargetColumn - nvarchar(50)
                         @Key , -- Key - nvarchar(50)
                         [RFOKey] , -- SourceValue - nvarchar(50)
                         [HybrisKey]
@@ -367,7 +376,9 @@ AS
       
 			
 -- MixNMatch
-        SET @message = ' STEP: 7.initiating MIXMATCH Validation '
+        SET @message = CONCAT(' STEP: 7.initiating MIXMATCH Validation ',
+                              CHAR(10),
+                              '-----------------------------------------------')
         EXECUTE dbqa.uspPrintMessage @message
 
 
@@ -453,7 +464,7 @@ AS
                   Owner ,
                   Flag ,
                   SourceColumn ,
-                  TargetCoulumn ,
+                  TargetColumn ,
                   [Key] ,
                   SourceValue ,
                   TargetValue
@@ -463,7 +474,7 @@ AS
                         @Owner , -- Owner - nvarchar(50)
                         @ValidationType , -- Flag - nvarchar(10)
                         RFOKey , -- SourceColumn - nvarchar(50)
-                        HybrisKey , -- TargetCoulumn - nvarchar(50)
+                        HybrisKey , -- TargetColumn - nvarchar(50)
                         @Key , -- Key - nvarchar(50)
                         RFC , -- SourceValue - nvarchar(50)
                         HBC
@@ -476,7 +487,9 @@ AS
 --++++++++++++++++++++++++++++++++++++++++++++++++++
 		
       	
-        SET @message = ' STEP: 8.Removing Issues for END TO END Validaion  '
+        SET @message = CONCAT(' STEP: 8.Removing Issues for END TO END Validaion  ',
+                              CHAR(10),
+                              '-----------------------------------------------')
         EXECUTE dbqa.uspPrintMessage @message
 		
 
@@ -488,12 +501,6 @@ AS
         FROM    #RFO a
                 JOIN #MixMatch m ON m.RFOkey = a.RFOKey
 
-
-        DELETE  a
-        FROM    #RFO a
-                JOIN #DupRFO m ON m.RFOkey = a.RFOKey
-
-
         DELETE  a
         FROM    #Hybris a
                 JOIN #missing m ON m.HybrisKey = a.HybrisKey	
@@ -501,17 +508,38 @@ AS
         DELETE  a
         FROM    #Hybris a
                 JOIN #MixMatch m ON m.HybrisKey = a.HybrisKey	
-        DELETE  a
-        FROM    #Hybris a
-                JOIN #DupHybris m ON m.HybrisKey = a.HybrisKey	
 
-        DROP TABLE #missing
-        DROP TABLE #MixMatch
-        DROP TABLE #DupRFO
-        DROP TABLE #DupHybris
+        IF OBJECT_ID('tempdb.dbo.#DupRFO') IS NOT NULL
+            DELETE  a
+            FROM    #RFO a
+                    JOIN #DupRFO m ON m.RFOkey = a.RFOKey
+
+
+        IF OBJECT_ID('tempdb.dbo.#DupHybris') IS NOT NULL
+            DELETE  a
+            FROM    #Hybris a
+                    JOIN #DupHybris m ON m.HybrisKey = a.HybrisKey	
+
+
+        IF OBJECT_ID('tempdb.dbo.#Missing') IS NOT NULL
+            DROP TABLE #Missing
+        IF OBJECT_ID('tempdb.dbo.#MixMatch') IS NOT NULL
+            DROP TABLE #MixMatch
+        IF OBJECT_ID('tempdb.dbo.#DupHybris') IS NOT NULL
+            DROP TABLE #DupHybris
+        IF OBJECT_ID('tempdb.dbo.#DupRFO') IS NOT NULL
+            DROP TABLE #DupRFO
  
 
+        SET @sourceCount = 0
+        SELECT  @sourceCount = COUNT(RFOKey)
+        FROM    #RFO a
+                JOIN #Hybris b ON a.RFOKey = b.HybrisKey
 
+        SET @message = CONCAT(' STEP: 9. Total Record Counts for End to End = ',
+                              CAST(@SourceCount AS NVARCHAR(25)), CHAR(10),
+                              '-----------------------------------------------')
+        EXECUTE dbqa.uspPrintMessage @message
 
 
         IF OBJECT_ID('Tempdb.dbo.#Temp') IS NOT NULL
@@ -524,7 +552,6 @@ AS
                 AND [flag] IN ( 'c2c', 'ref', 'default' )
 
 
-		
 		
         DECLARE @MaxRow INT ,
             @RowNumber INT ,
@@ -541,15 +568,17 @@ AS
             );
 
 
+
+       
         SELECT  @MaxRow = MAX(RowNumber)
         FROM    #Temp
         IF ISNULL(@MaxRow, 0) > 0
             BEGIN
-        
-                SET @Message = 'STEP: 9.Validation Started For Columnt To Column with  total fields= '
-                    + CAST(@MaxRow AS NVARCHAR(20))              
+                SET @Message = CONCAT('STEP: 10. Validation Started For Columnt To Column with  total fields= ',
+                                      CAST(@MaxRow AS NVARCHAR(20)), CHAR(10),
+                                      '-----------------------------------------------')
+                   
                 EXECUTE dbqa.uspPrintMessage @message
-
 
 
                 SET @RowNumber = 1
@@ -567,8 +596,11 @@ AS
 
                         SET @Message = CONCAT('Column Validation Started For ',
                                               CAST(@RowNumber AS NVARCHAR(20)),
-                                              '. ', @TargetColumn)
+                                              '. ', @TargetColumn, CHAR(10),
+                                              '-----------------------------------------------')
+
                         EXECUTE dbqa.uspPrintMessage @message
+ 
 
                         INSERT  INTO @temp
                                 ( [key], SourceValue,--[SourceColumn]
@@ -582,8 +614,10 @@ AS
                             BEGIN
                                 SET @Message = CONCAT('Total IssueCount=',
                                                       CAST(@rowCounts AS NVARCHAR(12)),
-                                                      ' for ', @TargetColumn)
-
+                                                      ' for ', @TargetColumn,
+                                                      CHAR(10),
+                                                      '-----------------------------------------------')
+                                EXECUTE dbqa.uspPrintMessage @message
 
                                 INSERT  INTO dbqa.SourceTargetLog
                                         ( FlowTypes ,
@@ -594,7 +628,7 @@ AS
                                           comments ,
                                           ExecutionStatus
                                         )
-                                VALUES  ( @Owner , -- FlowTypes - nvarchar(50)
+                                VALUES  ( @Flows , -- FlowTypes - nvarchar(50)
                                           CONCAT(@Flag, '_EndToEnd') , -- ValidationTypes - nvarchar(50)
                                           @Owner , -- Owner - nvarchar(50)
                                           @rowCounts , -- SourceCount - int
@@ -609,18 +643,18 @@ AS
                                           [Owner] ,
                                           Flag ,
                                           SourceColumn ,
-                                          TargetCoulumn ,
+                                          TargetColumn ,
                                           [Key] ,
                                           SourceValue ,
                                           TargetValue
                                         )
                                         SELECT TOP 10
-                                                @Owner ,
+                                                @Flows ,
                                                 @Owner ,
                                                 CONCAT(@Flag, '_EndToEnd') ,
                                                 @SourceColumn ,
                                                 @TargetColumn ,
-                                                @Key ,
+                                                CONCAT(@Key, '=', [Key]) ,
                                                 SourceValue ,
                                                 TargetValue
                                         FROM    @temp
@@ -659,5 +693,3 @@ AS
 
             END
     END 
-
-
